requirements.txt
amqp==5.0.2
asgiref==3.3.1
billiard==3.6.3.0
build==0.1.0
celery==5.0.5
certifi==2020.12.5
cffi==1.14.4
chardet==4.0.0
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.1.6
cryptography==3.3.1
defusedxml==0.7.0rc1
Django==3.1.4
django-3-jet==1.0.8
django-celery-beat==2.1.0
django-cors-headers==3.6.0
django-filter==2.4.0
django-flatpages-i18n==1.0.0
django-js-asset==1.2.2
django-modeltrans==0.5.0
django-modeltranslation==0.16.1
django-mptt==0.11.0
django-simple-history==2.12.0
django-timezone-field==4.0
djangorestframework==3.12.2
easy-thumbnails==2.7.1
environs==9.3.0
gunicorn==20.0.4
idna==2.10
importlib-metadata==3.3.0
instapi==0.1.4
kombu==5.0.2
Markdown==3.3.3
marshmallow==3.10.0
oauthlib==3.1.0
packaging==20.8
pep517==0.9.1
Pillow==8.0.1
prompt-toolkit==3.0.9
psycopg2-binary==2.8.6
pycparser==2.20
PyJWT==2.0.0
pyparsing==2.4.7
python-crontab==2.5.1
python-dateutil==2.8.1
python-dotenv==0.15.0
python-slugify==4.0.1
python3-openid==3.2.0
pytz==2020.4
redis==3.5.3
redis-tools==0.0.1
requests==2.25.1
requests-oauthlib==1.3.0
six==1.15.0
social-auth-app-django==4.0.0
social-auth-core==3.3.3
sqlparse==0.4.1
text-unidecode==1.3
toml==0.10.2
typing-extensions==3.7.4.3
urllib3==1.26.2
vine==5.0.0
wcwidth==0.2.5
zipp==3.4.0
